requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
google-cloud-storage
google-cloud-firestore
marqo
supabase